no need for this instance variable.
Since onboard_judgings and participants have a 1 to 1 relationship,
one can simply call participant.onboard_judging in the view. As it was,
the particular onboard_judging was being selected as if it belonged
to a hash with the keys being participant_ids. In fact, it behaved
more like an array, being an ActiveRecord relation. Thus, the correct
score would only be displayed if the participant_ids went completely in
order.