Remove config/secrets
The only "secret" we have is the SECRET_KEY_BASE. In the absence of this
file, dev and test will make one for us, prod will still use
`ENV["SECRET_KEY_BASE"]` if it's defined.

For now, I'm keeping the `ENV` approach in production: If I end up
getting docker/kamal to work then ENV variables are way easier.